+++ Makefile.am
@@ -240,8 +240,7 @@ ML_FILES += \
     ml/Config.cc \
     ml/dlib/dlib/all/source.cpp \
     ml/json/single_include/nlohmann/json.hpp \
-    ml/nml.h \
-    ml/nml.cc \
+    ml/ml-private.h \
     ml/ml.cc \
     $(NULL)
